Tag and ExcludeTag not an array when passed from VSTS (#93)
* Addings tests for the new $null check

VSTS passes in $null when a field is empty, Main code was updated to include a check to handle this but needed a test adding as well.

* Splitting Tag and ExcludeTag on ','

VSTS passes in a single string as it doesn't handle arrays correctly. This now splits on commas and strips quotes.

Adding tests to check for this behaviour as well to ensure it doesn't get broken in future.