Update HelperModule.psm1 to fix #18
When PowershellGallery is not reachable (e.g. on an airgapped build server), the Find-Module command writes an error message stating that it cannot find the given module (in this case, Pester). This error message caused the Pester extension task to fail when used in an Azure DevOps pipeline, even though the tests would be executed as expected.

To fix this, we set errors from Find-Module to be terminating errors so they can be handled by the try/catch block and fall back to the bundled Pester module.